extracts Settings to insomnia-common package 1
