+++ packages/insomnia-app/app/common/hotkeys-listener.ts
@@ -1,5 +1,7 @@
+import { KeyBindings, KeyCombination } from 'insomnia-common';
+
 import * as models from '../models';
-import type { HotKeyDefinition, KeyBindings, KeyCombination } from './hotkeys';
+import type { HotKeyDefinition } from './hotkeys';
 import { areSameKeyCombinations, getPlatformKeyCombinations } from './hotkeys';
 
 const _pressedHotKey = (event: KeyboardEvent, bindings: KeyBindings) => {
