+++ packages/insomnia-app/app/common/hotkeys.ts
@@ -1,3 +1,5 @@
+import { HotKeyRegistry, KeyBindings, KeyCombination } from 'insomnia-common';
+
 import { ALT_SYM, CTRL_SYM, isMac, META_SYM, SHIFT_SYM } from './constants';
 import { keyboardKeys } from './keyboard-keys';
 import { strings } from './strings';
@@ -10,32 +12,6 @@ export interface HotKeyDefinition {
   id: string;
   description: string;
 }
-
-/**
- * The combination of key presses that will activate a hotkey if pressed.
- */
-export interface KeyCombination {
-  ctrl: boolean;
-  alt: boolean;
-  shift: boolean;
-  meta: boolean;
-  keyCode: number;
-}
-
-/**
- * The collection of a hotkey's key combinations for each platforms.
- */
-export interface KeyBindings {
-  macKeys: KeyCombination[];
-  // The key combinations for both Windows and Linux.
-  winLinuxKeys: KeyCombination[];
-}
-
-/**
- * The collection of defined hotkeys.
- * The registry maps a hotkey by its reference id to its key bindings.
- */
-export type HotKeyRegistry = Record<string, KeyBindings>;
 
 function defineHotKey(id: string, description: string): HotKeyDefinition {
   return {
