+++ packages/insomnia-app/app/models/settings.ts
@@ -1,73 +1,13 @@
-import type { HttpVersion } from '../common/constants';
+import { HttpVersions, Settings as BaseSettings } from 'insomnia-common';
 import {
   getAppDefaultDarkTheme,
   getAppDefaultLightTheme,
   getAppDefaultTheme,
-  HttpVersions,
   UPDATE_CHANNEL_STABLE,
 } from '../common/constants';
 import { database as db } from '../common/database';
 import * as hotkeys from '../common/hotkeys';
 import type { BaseModel } from './index';
-
-export interface PluginConfig {
-  disabled: boolean;
-}
-
-export type PluginConfigMap = Record<string, PluginConfig>;
-
-export interface BaseSettings {
-  autoHideMenuBar: boolean;
-  autocompleteDelay: number;
-  deviceId: string | null;
-  disableHtmlPreviewJs: boolean;
-  disableResponsePreviewLinks: boolean;
-  disableUpdateNotification: boolean;
-  editorFontSize: number;
-  editorIndentSize: number;
-  editorIndentWithTabs: boolean;
-  editorKeyMap: string;
-  editorLineWrapping: boolean;
-  enableAnalytics: boolean;
-  environmentHighlightColorStyle: string;
-  filterResponsesByEnv: boolean;
-  followRedirects: boolean;
-  clearOAuth2SessionOnRestart: boolean;
-  fontInterface: string | null;
-  fontMonospace: string | null;
-  fontSize: number;
-  fontVariantLigatures: boolean;
-  forceVerticalLayout: boolean;
-  hotKeyRegistry: hotkeys.HotKeyRegistry;
-  httpProxy: string;
-  httpsProxy: string;
-  lineWrapping?: boolean;
-  maxHistoryResponses: number;
-  maxRedirects: number;
-  maxTimelineDataSizeKB: number;
-  noProxy: string;
-  nunjucksPowerUserMode: boolean;
-  pluginConfig: PluginConfigMap;
-  pluginPath: string;
-  preferredHttpVersion: HttpVersion;
-  proxyEnabled: boolean;
-  showPasswords: boolean;
-  theme: string;
-  autoDetectColorScheme: boolean;
-  lightTheme: string;
-  darkTheme: string;
-  timeout: number;
-  updateAutomatically: boolean;
-  updateChannel: string;
-  useBulkHeaderEditor: boolean;
-  useBulkParametersEditor: boolean;
-  validateAuthSSL: boolean;
-  validateSSL: boolean;
-  hasPromptedToMigrateFromDesigner: boolean;
-  hasPromptedOnboarding: boolean;
-  hasPromptedAnalytics: boolean;
-  isVariableUncovered?: boolean;
-}
 
 export type Settings = BaseModel & BaseSettings;
 export const name = 'Settings';
