+++ packages/insomnia-app/app/ui/components/hotkey.tsx
@@ -1,8 +1,8 @@
 import classnames from 'classnames';
+import { KeyBindings, KeyCombination } from 'insomnia-common';
 import React, { FC, memo } from 'react';
 
 import { isMac } from '../../common/constants';
-import type { KeyBindings, KeyCombination } from '../../common/hotkeys';
 import { constructKeyCombinationDisplay, getPlatformKeyCombinations } from '../../common/hotkeys';
 
 interface Props {
