+++ packages/insomnia-app/app/ui/components/panes/placeholder-request-pane.tsx
@@ -1,15 +1,15 @@
+import { HotKeyRegistry } from 'insomnia-common';
 import React, { FC, useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { hotKeyRefs } from '../../../common/hotkeys';
-import * as hotkeys from '../../../common/hotkeys';
 import { importFile } from '../../redux/modules/import';
 import { selectActiveWorkspace } from '../../redux/selectors';
 import { Hotkey } from '../hotkey';
 import { Pane, PaneBody, PaneHeader } from './pane';
 
 interface Props {
-  hotKeyRegistry: hotkeys.HotKeyRegistry;
+  hotKeyRegistry: HotKeyRegistry;
   handleCreateRequest: () => void;
 }
