+++ packages/insomnia-app/app/ui/components/panes/response-pane.tsx
@@ -2,6 +2,7 @@ import { autoBindMethodsForReact } from 'class-autobind-decorator';
 import classnames from 'classnames';
 import { clipboard, remote } from 'electron';
 import fs from 'fs';
+import { HotKeyRegistry } from 'insomnia-common';
 import { json as jsonPrettify } from 'insomnia-prettify';
 import mime from 'mime-types';
 import React, { PureComponent } from 'react';
@@ -9,7 +10,6 @@ import { Tab, TabList, TabPanel, Tabs } from 'react-tabs';
 
 import { AUTOBIND_CFG, PREVIEW_MODE_SOURCE } from '../../../common/constants';
 import { exportHarCurrentRequest } from '../../../common/har';
-import type { HotKeyRegistry } from '../../../common/hotkeys';
 import { getSetCookieHeaders } from '../../../common/misc';
 import * as models from '../../../models';
 import type { Environment } from '../../../models/environment';
