+++ packages/insomnia-app/app/ui/components/settings/shortcuts.tsx
@@ -1,13 +1,14 @@
 import { autoBindMethodsForReact } from 'class-autobind-decorator';
+import { HotKeyRegistry, KeyCombination } from 'insomnia-common';
 import React, { PureComponent } from 'react';
 
 import { AUTOBIND_CFG } from '../../../common/constants';
-import type { HotKeyDefinition, HotKeyRegistry, KeyCombination } from '../../../common/hotkeys';
 import {
   areKeyBindingsSameAsDefault,
   areSameKeyCombinations,
   constructKeyCombinationDisplay,
   getPlatformKeyCombinations,
+  HotKeyDefinition,
   hotKeyRefs,
   newDefaultKeyBindings,
   newDefaultRegistry,
