+++ packages/insomnia-app/app/ui/components/sidebar/sidebar-request-group-row.tsx
@@ -1,5 +1,6 @@
 import { autoBindMethodsForReact } from 'class-autobind-decorator';
 import classnames from 'classnames';
+import { HotKeyRegistry } from 'insomnia-common';
 import React, { PureComponent } from 'react';
 import { PropsWithChildren } from 'react';
 import { createRef } from 'react';
@@ -7,7 +8,6 @@ import { DragSource, DragSourceSpec, DropTarget, DropTargetMonitor, DropTargetSp
 import { connect } from 'react-redux';
 
 import { AUTOBIND_CFG } from '../../../common/constants';
-import { HotKeyRegistry } from '../../../common/hotkeys';
 import * as misc from '../../../common/misc';
 import { HandleRender } from '../../../common/render';
 import { RequestGroup } from '../../../models/request-group';
