+++ packages/insomnia-common/.eslintignore
@@ -1,0 +1,1 @@
+dist
