+++ packages/insomnia-common/.eslintrc.js
@@ -1,0 +1,4 @@
+/** @type { import('eslint').Linter.Config } */
+module.exports = {
+  extends: '../../.eslintrc.js',
+};
