+++ packages/insomnia-common/jest.config.js
@@ -1,0 +1,9 @@
+/** @type { import('@jest/types').Config.InitialOptions } */
+module.exports = {
+  preset: '../../jest-preset.js',
+  globals: {
+    'ts-jest': {
+      isolatedModules: false,
+    },
+  },
+};
