+++ packages/insomnia-common/src/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './hotkeys';
+export * from './settings';
