+++ packages/insomnia-common/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "./tsconfig.build.json",
+  "compilerOptions": {
+    "composite": false,
+    "rootDir": ".",
+  },
+  "include": [
+    "src",
+    "jest.config.js",
+    ".eslintrc.js",
+  ],
+  "exclude": [
+    "dist",
+  ],
+}
